feat(abi)!: pre-1.0 ABI migration of core structs
- HostInterface: remove never-invoked call_guest_method placeholder
  (152 -> 144 bytes, 17 fn ptrs, get_extension @136)
- RuntimeConfig 24 bytes: on_reload now fn(*mut c_void, ReloadPhase),
  new on_reload_user_data field for host callback context
- HostContractInterface 80 bytes: new registrant-owned user_data @40
  so factories need no static state
- RuntimeInterface.load_bundle takes StringView (no more c_char)
- polyplug_host_alloc/free no longer #[no_mangle] - allocation flows
  exclusively through HostInterface.alloc/free fields
- AbiError.code is u32 (defensive against untrusted plugin returns);
  message is static/runtime-owned, receiver never frees
- get_extension slot + extension id hashing for the extension system
- emitters: fix name:type splitter corrupting :: paths, Array-returning
  fn-ptr typedefs, JS layout-test missing assert import

+++ crates/polyplug_abi/src/lib.rs
@@ -60,6 +60,6 @@ mod tests {
     fn test_abi_error_ok() {
         let e: AbiError = AbiError::ok();
         assert!(e.is_ok());
-        assert_eq!(e.code, AbiErrorCode::Ok);
+        assert_eq!(e.code, AbiErrorCode::Ok as u32);
     }
 }
